Fix bug with added users removal
When a user was added and a re-render of UserList triggered, the added users would be removed. This is because the API is not updating with the added users. The solution has been to only set users from the API when users is empyt, so setUsers(data) only runs in the first re-render